refactor utils into Path trait
